fix hardcodeing of gfx908.  Use OARCH

+++ trunk/parallel_test/parallel_test.sh
@@ -59,14 +59,17 @@ echo "C RETURN CODE IS: $rc"
 
 echo "CONVERTING temp bc files to ll.  See files host_c.ll, device_c.ll"
 $TRUNK/bin/llvm-dis main-host-x86_64-unknown-linux-gnu.bc -o host_c.ll
-$TRUNK/bin/llvm-dis main-openmp-amdgcn-amd-amdhsa-gfx908.tmp.bc -o device_c.ll
-echo "-----------------------------------------------------"
-echo "===> HOST function defs and calls in tmpc/host_c.ll"
-grep "define\|call" host_c.ll | grep -v requires | grep -v nocallback | tee host_calls.txt
-echo "-----------------------------------------------------"
-echo "===> DEVICE function defs and calls in tmpc/device_c.ll"
-grep "define\|call" device_c.ll | grep -v nocallback | tee device_calls.txt
-echo "-----------------------------------------------------"
+_bcfile="main-openmp-amdgcn-amd-amdhsa-$OARCH.tmp.bc"
+if [ -f $_bcfile ] ; then
+   $TRUNK/bin/llvm-dis $_bcfile -o device_c.ll
+   echo "-----------------------------------------------------"
+   echo "===> HOST function defs and calls in tmpc/host_c.ll"
+   grep "define\|call" host_c.ll | grep -v requires | grep -v nocallback | tee host_calls.txt
+   echo "-----------------------------------------------------"
+   echo "===> DEVICE function defs and calls in tmpc/device_c.ll"
+   grep "define\|call" device_c.ll | grep -v nocallback | tee device_calls.txt
+   echo "-----------------------------------------------------"
+fi
 echo
 tmpf="tmpf"
 echo
@@ -103,8 +106,9 @@ if [ -f main-host-x86_64-unknown-linux-gnu.bc ] ; then
    echo "===> HOST function defs and calls in tmpf/host_f.ll"
    grep "define\|call" host_f.ll | grep -v requires | grep -v nocallback | tee host_calls.txt
 fi
-if [ -f main-openmp-amdgcn-amd-amdhsa-gfx908.tmp.bc ] ; then 
-   $TRUNK/bin/llvm-dis main-openmp-amdgcn-amd-amdhsa-gfx908.tmp.bc -o device_f.ll
+_bcfile="main-openmp-amdgcn-amd-amdhsa-$OARCH.tmp.bc"
+if [ -f $_bcfile ] ; then
+   $TRUNK/bin/llvm-dis $_bcfile -o device_f.ll
    echo "-----------------------------------------------------"
    echo "===> DEVICE function defs and calls in tmpf/device_f.ll"
    grep "define\|call" device_f.ll | grep -v nocallback
